major changes related to transforms
