+++ src/geome/adata2data/__init__.py
@@ -1,0 +1,3 @@
+from .anndata2data import AnnData2Data
+from .default import AnnData2DataDefault
+from .by_category import AnnData2DataByCategory
