+++ src/geome/iterables/__init__.py
@@ -1,0 +1,2 @@
+from .adata2iterable import AnnData2Iterable,
+from .to_category_iterable import ToCategoryIterable
