+++ src/geome/transforms/__init__.py
@@ -1,0 +1,3 @@
+from .compose import Compose
+from .to_array import ToArray
+from .add_design_matrix import AddDesignMatrix
